requirements.txt
Django==2.1.7
django-stubs==0.12.1
djangorestframework==3.9.3
djangorestframework-simplejwt==4.3.0
djangorestframework-stubs==0.4.2
mypy==0.701
mypy-extensions==0.4.1
mysqlclient==1.4.2.post1
Pillow==6.0.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2019.1
typed-ast==1.3.5
typing-extensions==3.7.2
uuid==1.30
